requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.3.2  # https://github.com/redis/hiredis-py
celery==5.3.6  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower
boto3~=1.34.42
botocore~=1.34.42
requests~=2.31.0
responses~=0.25.0
sorl-thumbnail==12.10.0
pytest-xdist~=3.6.1

# Django
# ------------------------------------------------------------------------------
django==4.2.10  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.4.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.61.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2023.10  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-simple-history~=3.4.0
django-cachalot==2.6.2
django-enumchoicefield==3.0.1
django-silk==5.1.0

# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.3.1  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.27.1  # https://github.com/tfranzel/drf-spectacular
# Wagtail
wagtail==6.0.2  # https://github.com/wagtail/wagtail

# ffmpeg
ffmpeg~=1.4
ffmpeg-python==0.2.0
